+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This file is used to list changes made in each version of the openstack-ops-data
 ## 10.0.0
 * Upgrading to Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5
+* Switching mysql server recipe to use resource and set password
+  directly
 
 ## 9.2.0
 * Update database and mysql dependency